Coral-SZ/requirements.txt
absl-py
appdirs
astunparse
attrs
audioread
cachetools
certifi
cffi
charset-normalizer
cycler
Cython
dataclasses
decorator
dill
dm-tree
etils
fire
flatbuffers
gast
gin-config
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
h5py
httplib2
idna
importlib-metadata
importlib-resources
joblib
kaggle
keras
Keras-Preprocessing
kiwisolver
libclang
librosa
llvmlite
lxml
Markdown
matplotlib
neural-structured-learning
numba
numpy
oauthlib
opencv-python-headless
opt-einsum
packaging
pandas
Pillow
pooch
promise
proto-plus
protobuf
psutil
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules
pybind11
pycocotools
pycparser
pyparsing
python-dateutil
python-slugify
pytz
PyYAML
requests
requests-oauthlib
resampy
rsa
scann
scikit-learn
scipy
sentencepiece
six
sounddevice
SoundFile
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-addons
tensorflow-datasets
tensorflow-hub
tensorflow-io-gcs-filesystem
tensorflow-metadata
tensorflow-model-optimization
tensorflowjs
termcolor
text-unidecode
tf-estimator-nightly
tf-models-official
tf-slim
tflite-model-maker
tflite-support
threadpoolctl
toml
tqdm
typeguard
typing_extensions
uritemplate
urllib3
Werkzeug
wrapt
zipp==3.8.0
PySide6==6.2.4
mendel-development-tool==1.5.2